_여분_현대오토에버_수어톡/src/sueotalk/pyproject.toml
[project]
name = "sueotalk"
version = "0.1.0"
description = "수어톡 (SuEoTalk) — 한국수어 ↔ 한국어 양방향 통역 Streamlit PoC"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.36",
    "pillow>=10.0",
    "numpy>=1.26",
    "httpx>=0.27",
]

[project.optional-dependencies]
vision = [
    "mediapipe>=0.10",
    "opencv-python-headless>=4.9",
]
stt = [
    "faster-whisper>=1.0",
]
tts = [
    "pyttsx3>=2.90",
]
dev = [
    "ruff>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["app.py", "holistic.py", "ksl_dictionary.py", "stt_runner.py", "avatar.py", "presets.py", "data"]

[tool.ruff]
line-length = 100
target-version = "py311"
